+++ src/core.ts
@@ -94,6 +94,11 @@ export interface Options {
   /** Determines how to treat inputs and outputs. @default ProcessingMode.CLONE_OFF */
   readonly processingMode?: ProcessingMode;
   /**
+   * Enforces strict MongoDB compatibilty. See readme for differences. @default true.
+   * When disabled, the $elemMatch projection operator returns all matching nested documents instead of only the first.
+   */
+  readonly useStrictMode?: boolean;
+  /**
    * Enables or disables custom script execution.
    * When disabled, you cannot use operations that execute custom code, such as the $where, $accumulator, and $function.
    * @default true
@@ -125,6 +130,7 @@ export function initOptions(options?: Options): Options {
   return Object.freeze({
     idKey: "_id",
     scriptEnabled: true,
+    useStrictMode: true,
     processingMode: ProcessingMode.CLONE_OFF,
     currentTimestamp: Date.now(),
     ...options,